requirements.txt
anyio==4.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
instagrapi==1.17.10
jiter==0.12.0
openai==2.7.2
pillow==12.0.0
pycryptodomex==3.17
pydantic==1.10.7
PySocks==1.7.1
python-dotenv==1.2.1
requests==2.32.5
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.5.0